[REFACTOR] move orchestration to application services
